relax the group dependency
